Accum.dtype + required Init.dtype for typed accumulators
Loop IR: Accum gains an optional dtype field (None = derive from value,
preserves today's behavior). Kernel IR: Init.dtype is keyword-only and
required, since placing an Init is the freeze point — every pass that
emits an Init must commit to a concrete dtype.

The Init-placement pass (000_place_inits) now decides the accumulator
dtype at freeze time and stamps both the Init and every matching Accum
with the same DataType. Today's policy is "always F32"; the follow-up
that promotes fp16 reductions to F32 accumulator plugs into _decide_dtype.

Init.render + Loop/StridedLoop's implicit init now declare in the
selected dtype (float vs __half) and use a dtype-aware identity
(__float2half(0.0f) for fp16). Accum.render inserts a __half2float /
__float2half conversion on `value` only when its dtype disagrees with
the accumulator's. RenderCtx.ssa_dtypes tracks accumulator dtype the
same way it tracks Load/Assign results.

Graph serialization round-trip: added DataType + np.dtype to the
Stmt-repr eval scope so Init/Accum reprs survive to_dict/from_dict.

render_kernelop falls back to KernelOp.input_tensors/output_tensors
when the caller doesn't pass a tensors= map (tests construct bare
KernelOps with input_tensors set).

New test: fp32-accumulator-over-fp16-loads renders the expected
`float acc = 0.0f; ... acc += __half2float(a); ...` pattern.

+++ deplodock/compiler/ir/stmt/passes.py
@@ -70,12 +70,12 @@ def _(s: Assign, rename: Rename, sigma: Sigma, axis_fn: AxisFn) -> Stmt:
 
 @rewrite.register
 def _(s: Accum, rename: Rename, sigma: Sigma, axis_fn: AxisFn) -> Stmt:
-    return Accum(name=rename(s.name), value=rename(s.value), op=s.op)
+    return Accum(name=rename(s.name), value=rename(s.value), op=s.op, dtype=s.dtype)
 
 
 @rewrite.register
 def _(s: Init, rename: Rename, sigma: Sigma, axis_fn: AxisFn) -> Stmt:
-    return Init(name=rename(s.name), op=s.op)
+    return Init(name=rename(s.name), op=s.op, dtype=s.dtype)
 
 
 @rewrite.register
